Move 'collate' to 'editing' module and add test
